Testing out filters

+++ Lab2/app/src/main/java/lab2_202_10/uwaterloo/ca/lab2/gesture/filter/Filter.java
@@ -2,14 +2,14 @@ package lab2_202_10.uwaterloo.ca.lab2.gesture.filter;
 
 public abstract class Filter {
 
-    public double[] filter(double[] vector) {
+    public float[] filter(float[] vector) {
         if (vector == null) {
             return null;
         }
         return filterAlgorithm(vector);
     }
 
-    abstract public double[] filterAlgorithm(double[] vector);
+    abstract public float[] filterAlgorithm(float[] vector);
 
     abstract public void reset();
